+++ concretedesignpy/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.7.3'
+__version__ = '0.7.4'
 
 from .general import area_diam, steel_area, area_ratio
 from .mgt import generateFIBERMGT